Add support for version command
- Also refactor the version code out into top-level version pkg to match
how the control-plane is factored
