+++ cli/commands.go
@@ -6,6 +6,8 @@ import (
 	"github.com/hashicorp/consul-k8s/cli/cmd/common"
 	"github.com/hashicorp/consul-k8s/cli/cmd/install"
 	"github.com/hashicorp/consul-k8s/cli/cmd/uninstall"
+	cmdversion "github.com/hashicorp/consul-k8s/cli/cmd/version"
+	"github.com/hashicorp/consul-k8s/cli/version"
 	"github.com/hashicorp/go-hclog"
 	"github.com/mitchellh/cli"
 )
@@ -26,6 +28,12 @@ func initializeCommands(ctx context.Context, log hclog.Logger) (*common.BaseComm
 		"uninstall": func() (cli.Command, error) {
 			return &uninstall.Command{
 				BaseCommand: baseCommand,
+			}, nil
+		},
+		"version": func() (cli.Command, error) {
+			return &cmdversion.Command{
+				BaseCommand: baseCommand,
+				Version:     version.GetHumanVersion(),
 			}, nil
 		},
 	}
