+++ cli/main.go
@@ -6,7 +6,7 @@ import (
 	"os/signal"
 	"syscall"
 
-	"github.com/hashicorp/consul-k8s/cli/cmd/version"
+	"github.com/hashicorp/consul-k8s/cli/version"
 	"github.com/hashicorp/go-hclog"
 	"github.com/mitchellh/cli"
 )
